refactor(lobster): merge 4 approval-prep steps into request_spec_approval
Collapses prepare_topic_approval, ensure_topic_slot_available,
finalize_review_cycle, and compact_approval_preview into a single
lobster_request_spec_approval.py step. The four phases run sequentially
within one process so finalize always commits state before the lobster
approval gate fires — no change in observable behaviour.

Also fixes a per-topic vs global blocking inconsistency: the merged step
uses per-topic locking throughout so different topics can run concurrently.
